requirements.txt
langchain
langchain-community
langchain-google-genai
langchain-core
langchain-text-splitters
faiss-cpu
pypdf
python-dotenv
fastapi
uvicorn[standard]
python-multipart
